refactor: UiModel mapping을 companion object에서 수행하도록 수정
