--- app/src/main/java/woowacourse/movie/feature/complete/ui/ReservationCompleteEntity.kt
@@ -1,9 +1,0 @@
-package woowacourse.movie.feature.complete.ui
-
-import woowacourse.movie.model.Ticket
-import woowacourse.movie.model.data.dto.Movie
-
-class ReservationCompleteEntity(
-    val movie: Movie,
-    val ticket: Ticket,
-)
