+++ app/src/main/java/woowacourse/movie/feature/seat/ui/SeatSelectMovieUiModel.kt
@@ -1,3 +1,11 @@
 package woowacourse.movie.feature.seat.ui
 
-class SeatSelectMovieUiModel(val titleMessage: String)
+import woowacourse.movie.model.data.dto.Movie
+
+class SeatSelectMovieUiModel(val titleMessage: String) {
+    companion object {
+        fun from(movie: Movie): SeatSelectMovieUiModel {
+            return SeatSelectMovieUiModel(movie.title)
+        }
+    }
+}
