--- app/src/main/java/woowacourse/movie/feature/seat/ui/SeatSelectMovieUiModelMapper.kt
@@ -1,7 +1,0 @@
-package woowacourse.movie.feature.seat.ui
-
-import woowacourse.movie.model.data.dto.Movie
-
-fun Movie.toSeatSelectMovieUiModel(): SeatSelectMovieUiModel {
-    return SeatSelectMovieUiModel(title)
-}
